code/requirements.txt
sklearn~=0.0
jsonlines~=1.2.0
pytorch_pretrained_bert
transformers~=3.5.1
fasttext~=0.9.2
wordsegment~=1.3.1
seqeval~=1.2.2
google-cloud-translate==2.0.1
pandas~=1.1.0
numpy~=1.18.1
torch~=1.7.0
tqdm~=4.43.0
twitter-text-python~=1.1.1
typing~=3.7.4.3
testtools~=2.4.0
six~=1.14.0
scipy~=1.4.1
argparse~=1.4.0
scikit-learn~=0.23.2
dataclasses~=0.6
requests~=2.24.0
matplotlib~=3.3.1
nltk~=3.5
tweepy~=3.9.0
Unidecode~=1.1.1
seaborn~=0.10.0